Add full support for Huggingface BitNet

+++ python/t_mac/utils.py
@@ -20,7 +20,7 @@ _device_kwargs = {
         # "remote_kwargs": None,
         "cc_opts": ["-O3", "-std=c++17", "-mcpu=apple-m2", "-mllvm", "-inline-threshold=10000"],
         "out_dtype": "float16",
-        "aggregation_dtype": "int16",
+        "aggregation_dtype": "int32",
     },
     "android": {
         "target": "llvm -device=arm_cpu -mtriple=aarch64-linux-gnu -mattr=+v8.2a,+fullfp16,+fp-armv8,+neon",
@@ -37,7 +37,7 @@ _device_kwargs = {
         },
         "cc_opts": ["-O3", "-march=armv8.2a+fp16"],
         "out_dtype": "float16",
-        "aggregation_dtype": "int16",
+        "aggregation_dtype": "int32",
     },
     "intel_win": {
         "target": "llvm -mtriple=x86_64-pc-windows-msvc -mcpu=core-avx2",
@@ -60,7 +60,7 @@ _device_kwargs = {
         "remote_kwargs": None,
         "cc_opts": ["-O3", "-std=c++17", "-march=armv8.2a+fp16", "-mllvm", "-inline-threshold=10000"],
         "out_dtype": "float16",
-        "aggregation_dtype": "int16",
+        "aggregation_dtype": "int32",
     },
 }
